+++ src/de/uni_koblenz/jgralab/greql2/evaluator/vertexeval/BagComprehensionEvaluator.java
@@ -132,7 +132,7 @@ public class BagComprehensionEvaluator extends VertexEvaluator {
 	}
 
 	@Override
-	public int calculateEstimatedCardinality(GraphSize graphSize) {
+	public long calculateEstimatedCardinality(GraphSize graphSize) {
 		return greqlEvaluator.getCostModel()
 				.calculateCardinalityBagComprehension(this, graphSize);
 	}
